お気に入り画面のホテルの金額がDoubleになっていたのでIntに変更、一行あたりのデザインを整えた

+++ app/src/main/java/com/kuro/trip_favo/data/api/RakutenHotelResult.kt
@@ -38,7 +38,7 @@ data class HotelBasicInfo(
             address1,
             address2,
             reviewAverage,
-            hotelMinCharge.toDouble()
+            hotelMinCharge
         )
     }
 }
